+++ src/program/protocols/union/common.rs
@@ -8,6 +8,7 @@ use protocol_builder::{
     types::{input::SpendMode, InputArgs, OutputType},
 };
 use sha2::{Digest, Sha256};
+use tracing::info;
 use uuid::Uuid;
 
 use crate::{
@@ -15,8 +16,8 @@ use crate::{
     program::{
         participant::ParticipantRole,
         protocols::union::types::{
-            PenalizedMember, StreamSettings, UnionSettings, GLOBAL_SETTINGS_UUID, OP_CLAIM_GATE,
-            PAIRWISE_DISPUTE_KEY, WT_CLAIM_GATE,
+            PenalizedMember, StreamSettings, UnionSettings, GLOBAL_SETTINGS_UUID, MY_IDX,
+            OP_CLAIM_GATE, PAIRWISE_DISPUTE_KEY, WT_CLAIM_GATE,
         },
         variables::{PartialUtxo, VariableTypes},
     },
@@ -575,9 +576,15 @@ pub fn save_penalized_member(
     committee_id: Uuid,
     data: &PenalizedMember,
 ) -> Result<(), BitVMXError> {
+    let name = data.storage_name();
+    info!(
+        "Updating penalized member data in storage: {}. Data: {:?}",
+        name, data
+    );
+
     context.globals.set_var(
         &committee_id,
-        &data.storage_name(),
+        &name,
         VariableTypes::String(serde_json::to_string(data)?),
     )?;
     Ok(())
@@ -597,4 +604,21 @@ pub fn load_penalized_member(
 
     let data: PenalizedMember = serde_json::from_str(&var.string()?)?;
     Ok(Some(data))
+}
+
+pub fn set_my_idx(context: &ProgramContext, pid: Uuid, my_idx: usize) -> Result<(), BitVMXError> {
+    context
+        .globals
+        .set_var(&pid, MY_IDX, VariableTypes::Number(my_idx as u32))?;
+    Ok(())
+}
+
+pub fn get_my_idx(context: &ProgramContext, pid: Uuid) -> Result<usize, BitVMXError> {
+    match context.globals.get_var(&pid, MY_IDX)? {
+        Some(var) => Ok(var.number()? as usize),
+        None => Err(BitVMXError::InvalidParameter(format!(
+            "My index not found for protocol {}",
+            pid
+        ))),
+    }
 }