+++ src/program/protocols/union/dispute_core.rs
@@ -13,9 +13,9 @@ use crate::{
                     estimate_fee, extract_double_index, extract_index,
                     extract_index_from_claim_gate, get_accept_pegin_pid, get_dispatch_action,
                     get_dispute_channel_pid, get_dispute_core_pid, get_dispute_pair_key_name,
-                    get_full_penalization_pid, get_initial_deposit_output_type, get_stream_setting,
-                    indexed_name, load_penalized_member, load_union_settings, triple_indexed_name,
-                    InputSigningInfo, WinternitzData,
+                    get_full_penalization_pid, get_initial_deposit_output_type, get_my_idx,
+                    get_stream_setting, indexed_name, load_penalized_member, load_union_settings,
+                    set_my_idx, triple_indexed_name, InputSigningInfo, WinternitzData,
                 },
                 dispute_core_claim_gate::{
                     ClaimGateAction, CLAIM_GATE_INIT_STOPPER_COMMITTEE_LEAF,
@@ -75,6 +75,7 @@ const REVEAL_INPUT_TX_COMMITTEE_LEAF: usize = 1;
 
 // const WT_INIT_CHALLENGE_TX_COSIGN_LEAF: usize = 0;
 const WT_INIT_CHALLENGE_TX_TIMELOCK_LEAF: usize = 1;
+pub const WT_INIT_CHALLENGE_TX_COSIGN_DISABLER_LEAF: usize = 2;
 
 const WT_INIT_CHALLENGE_COSIGN_VOUT: u32 = 0;
 const WT_INIT_CHALLENGE_WT_STOPPER_VOUT: u32 = 3;
@@ -371,6 +372,7 @@ impl ProtocolHandler for DisputeCoreProtocol {
         context: &ProgramContext,
     ) -> Result<(), BitVMXError> {
         info!("Building DisputeCoreProtocol for program {}", self.ctx.id);
+        set_my_idx(context, self.ctx.id, self.ctx.my_idx)?;
         let dispute_core_data = self.dispute_core_data(context)?;
         self.validate_keys(&keys, context, dispute_core_data.committee_id)?;
 
@@ -1688,8 +1690,11 @@ impl DisputeCoreProtocol {
         role: ParticipantRole,
         pid: Uuid,
     ) -> Result<(), BitVMXError> {
+        // Need to load my_idx from storage because self.ctx.my_idx has my index on DRP
+        let my_idx = get_my_idx(program_context, self.ctx.id)?;
+
         info!(
-            id = self.ctx.my_idx,
+            id = my_idx,
             "Handling wins action {}. PID: {}", tx_name, pid
         );
 
@@ -1702,7 +1707,6 @@ impl DisputeCoreProtocol {
             .members
             .iter()
             .enumerate()
-            .filter(|(_, m)| m.role == ParticipantRole::Prover)
             .find(|(op_index, _)| get_dispute_channel_pid(committee_id, *op_index, wt_index) == pid)
             .map(|(op_index, _)| op_index);
 
@@ -1710,15 +1714,20 @@ impl DisputeCoreProtocol {
             Some(i) => i,
             None => {
                 error!(
-                    id = self.ctx.my_idx,
+                    id = my_idx,
                     "Could not find matching DisputeChannel program for PID: {}", pid
                 );
                 return Ok(());
             }
         };
 
+        info!(
+            id = my_idx,
+            "DisputeChannel operator index for PID {} is {}", pid, drp_op_index
+        );
+
         if role == ParticipantRole::Prover {
-            if self.ctx.my_idx == drp_op_index {
+            if my_idx == drp_op_index {
                 self.dispatch_claim_gate(
                     program_context,
                     ClaimGateAction::Start,
@@ -1727,7 +1736,7 @@ impl DisputeCoreProtocol {
                 )?;
             }
         } else if role == ParticipantRole::Verifier {
-            if self.is_my_dispute_core(program_context)? {
+            if my_idx == wt_index {
                 self.dispatch_claim_gate(
                     program_context,
                     ClaimGateAction::Start,
@@ -2025,15 +2034,15 @@ impl DisputeCoreProtocol {
                 Some(penalized_member) => {
                     info!(
                     id = self.ctx.my_idx,
-                    "Operator already penalized for member index: {}, skipping OP_COSIGN dispatch",
+                    "Watchtower already penalized for member index: {}, skipping OP_COSIGN dispatch",
                     data.member_index
                 );
 
                     self.dispatch(
                         context,
                         DisputeCoreTxType::PenalizationWatchtowerDisabler {
-                            wt_index: penalized_member.challenger_index,
-                            op_disabler_directory_index: penalized_member.member_index,
+                            wt_index: penalized_member.member_index,
+                            op_disabler_directory_index: penalized_member.challenger_index,
                             op_index,
                         },
                     )?;
@@ -2041,8 +2050,8 @@ impl DisputeCoreProtocol {
                     self.dispatch(
                         context,
                         DisputeCoreTxType::PenalizationWatchtowerCosignDisabler {
-                            wt_index: penalized_member.challenger_index,
-                            op_disabler_directory_index: penalized_member.member_index,
+                            wt_index: penalized_member.member_index,
+                            op_disabler_directory_index: penalized_member.challenger_index,
                             op_index,
                         },
                     )?;